Add locations check for bad ref and bad parent
