+++ pyproject.toml
@@ -48,4 +48,4 @@ pack-checker = "pack_checker.cli:main"
 [tool.flake8]
 doctests = true
 max-line-length = 120
-max-complexity = 39  # TODO: get this down to ~15
+max-complexity = 41  # TODO: get this down to ~15
